services/course-parser/requirements.txt
fastapi
uvicorn
pdf2image
requests
python-multipart
google-generativeai